Add separate service .class file back
